% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify-package.R
\docType{package}
\name{notify-package}
\alias{notify-package}
\title{notify: Notify}
\description{
Provides a drop-in functor interactive logging in pipe chains.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoFrhwld/notify}
  \item \url{https://jofrhwld.github.io/notify/}
  \item Report bugs at \url{https://github.com/JoFrhwld/notify/issues}
}

}
\author{
\strong{Maintainer}: Josef Fruehwald \email{JoFrhwld@gmail.com}

}
\keyword{internal}
