% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Notify}
\usage{
notify(expr, msg)
}
\arguments{
\item{expr}{The input expression}

\item{msg}{The logging message. If none is provided,
it will try to provide a reasonable message.}
}
\value{
The evaluated \code{expr}
}
\description{
Use \code{notify()} to insert interactive logging into
a sequence of functions chained by a base pipe (\verb{|>}).
Does not work with magrittr \verb{\%>\%}.
}
\examples{
mtcars |>
  notify("Using mtcars") |>
  cov() |>
  notify("Got full covariance matrix") |>
  colMeans() |>
  notify("Calculated column means")

# without explicit messages
mtcars |>
  notify() |>
  cov() |>
  notify() |>
  colMeans() |>
  notify()

}
